maf/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-framework-core==1.0.0b260212",
    "agent-framework-azure-ai==1.0.0b260212",
    "agent-framework-ag-ui==1.0.0b260212",
    "agent_framework.orchestrations==1.0.0b260212",
    "azure-ai-agents==1.2.0b5",
    "azure-ai-projects==2.0.0b3",
    "python-dotenv",
]

